Some initial pieces of the POC
We use gitea as SCM provider, so this includes some values for
deploying gitea on the IBM cloud cluster.
Repo can be cloned anonymously but anything else requires login.

Tekton triggers and pipeline are going to be used, in combination
with a new tool, "cdeventer", to react to webhook/events and
produce cdevents. We can use vanilla Tekton here so pretty much
just install the components for now.

The Tekton dashboard is handy for the demo, and it's deployed
behind an oauth-proxy to secure it via GitHub login.

cdeventer itself is a controller plus tekton resources.
The controller for now is added on the cluster using ko, and it's
not yet included in this commit.
The Tekton resources are developed in the POC and will be copied
over to cdeventer later on to make the more reusable.

The sink for CDEvents is a CloudEvent broker provided by Knative
with its in-memory channel implementation. Knative triggers
are used to define subscriptions and get CDEvents to the
interested parties.

The flow to produce CDEvent for now goes as follows:
- demo developer merges a PR in gitea
- webhook goes to an event listener in the tekton-ci namespace
- the tekton trigger/binding extracts the data required for the
  change CDEvent from the webhook headers and payload
- the tekton trigger template creates a "Run" i.e. a custom
  task run with the data for the CDEvent
- the cdeventer custom task reconciler reconciles the "Run".
  It uses the new go-sdk to build, validate and send a CDEvent
  to the CloudEvents broker

Now, to consume / display the CDEvent, the flows for now is:
- an all catching knative trigger gets the CDEvent to a different
  Tekton event-listener, el-events, in the default namespace
- *note* the knative broker must be set to 0 retries because of
  https://github.com/tektoncd/triggers/issues/1439
- A Tekton trigger matched on CDEvents runs a display pipeline
  that shows headers and payload of the CDEvent in the Tekton
  dashboard

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>